# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Products API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category information",
                "description": "Returns category data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Category id",
                        "required": true
                    },
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Category not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete a category",
                "description": "Soft delete a category",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Category id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/categories/{id}/translate": {
            "post": {
                "tags": [
                    "Categories Translations"
                ],
                "summary": "Post a new category translation",
                "description": "Create a new translation",
                "operationId": "addTranslation",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Category id",
                        "required": true
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Text",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Translation created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories Translations"
                ],
                "summary": "Delete a category translation",
                "description": "Soft delete a translation",
                "operationId": "removeTranslation",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Category id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product information",
                "description": "Returns product data",
                "operationId": "retrieve",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product id",
                        "required": true
                    },
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Product not found."
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Soft delete a product",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Patch a product",
                "description": "Update a product",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product id",
                        "required": true
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "available",
                        "required": false
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category Id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products information",
                "description": "Returns product data",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": false
                    },
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category Id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Servor Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Post a new product",
                "description": "Create a product",
                "operationId": "create",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "query",
                        "description": "Locale",
                        "required": true
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Description",
                        "required": true
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Category Id",
                        "required": true
                    },
                    {
                        "name": "store_id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Available",
                        "required": true
                    },
                    {
                        "name": "ht",
                        "in": "query",
                        "description": "HT",
                        "required": true
                    },
                    {
                        "name": "tva_rate",
                        "in": "query",
                        "description": "TVA rate",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Product created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/products/{id}/price": {
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Patch a product price",
                "description": "Update a product price",
                "operationId": "updatePrice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Product id",
                        "required": true
                    },
                    {
                        "name": "ht",
                        "in": "query",
                        "description": "HT",
                        "required": true
                    },
                    {
                        "name": "tva_rate",
                        "in": "query",
                        "description": "TVA rate",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}